Release 0.038
Bug fixes and feature changes

NOTE: not all is backwards compatible!

A revamp of the hidden _pp flag, now includes _info stuff to enable the listing
of conditions (e.g. versions of modules, config etc) for the non-packed and
packed copy respectively.

--defines didn't make it into the meta file

renamed some samples to better use .t as the preferred (perl code) test
extension

Addition of a postprocess mechanism similar to preprocess

The default/internal execmap is now not automatically used for safety reasons,
must be requested.